+++ db/migrate/20140611072144_add_permission_to_visualizations.rb
@@ -1,0 +1,9 @@
+Sequel.migration do
+  up do
+    add_column :visualizations, :permission_id, :uuid
+  end
+
+  down do
+    drop_column :visualizations, :permission_id
+  end
+end